Various minor fixes to make tutorials work and standardize

+++ src/main/java/me/neoblade298/neorogue/player/PlayerSessionData.java
@@ -1185,8 +1185,8 @@ public class PlayerSessionData extends MapViewer implements Comparable<PlayerSes
 	public void updateBoardLines() {
 		boardLines = new ArrayList<String>();
 		boardLines.add("§cHP§7: §f" + (int) health + "§7 / §f" + (int) maxHealth);
-		boardLines.add("§9MP§7: §f" + maxMana + " §7| §f" + df.format(manaRegen) + "/s");
-		boardLines.add("§aSP§7: §f" + maxStamina + " §7| §f" + df.format(staminaRegen) + "/s");
+		boardLines.add("§9MP§7: §f" + (int) maxMana + " §7| §f" + df.format(manaRegen) + "/s");
+		boardLines.add("§aSP§7: §f" + (int) maxStamina + " §7| §f" + df.format(staminaRegen) + "/s");
 		boardLines.add("§eCoins§7: §f" + coins);
 		s.updateSpectatorLines();
 		if (s.getParty().size() <= 1) return;
